requirements.txt
amqp==5.2.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.1.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap4==2024.1
cron-descriptor==1.4.5
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.4
django-after-response==0.2.2
django-browser-reload==1.12.1
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-ckeditor-5==0.2.13
django-cors-headers==4.3.1
django-countries==7.6.1
django-crispy-forms==2.3
django-decorator-include==3.0
django-easy-audit==1.3.6
django-etc==1.4.0
django-extensions==3.2.3
django-filter==24.2
django-hitcount==1.3.5
django-htmlmin==0.11.0
django-htmx==1.17.3
django-import-export==3.3.8
django-model-utils==5.0.0
django-multiselectfield==0.1.12
django-notifications-hq==1.8.3
django-storages==1.14.2
django-summernote==0.8.20.0
django-tenant-users==2.0.0
django-tenants==3.6.1
django-timezone-field==7.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docutils==0.21.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
html5lib==1.1
inflection==0.5.1
jsonfield==3.1.0
kombu==5.4.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.1
pillow==10.3.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
redis==5.0.3
setuptools==69.5.1
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
swapper==1.4.0
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.43.0
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
